warga rt rw perangkat desa pimpinan desa controller

+++ be/controllers/adminController/AdminController.js
@@ -122,12 +122,22 @@ exports.postWarga = async (req,res) => {
 exports.postRt = async (req, res) => {
     try {
         const { idUser } = req.params;
+        // mencari model warga berdasarkan idUser dan delete
+        const cekWarga = await WargaModel.findOne({user: idUser});
+        if (!cekWarga) {
+            throw new Error("User belum terdaftar sebagai warga");
+        }
+        await WargaModel.findOneAndDelete({user: idUser});
+
+        // mencari model rw berdasarkan idUser dan delete
+        const cekRw = await rwModel.findOne({user: idUser});
+        if (cekRw) {
+            await rwModel.findOneAndDelete({user: idUser});
+        }
+
         const dataUser = await userModel.findByIdAndUpdate(idUser, {
             role: 2 // (rt)
         }, { new: true });
-
-        // mencari model warga berdasarkan idUser dan delete
-        await WargaModel.findOneAndDelete({ user: idUser });
 
         const newRt = await rtModel.create({
             user: idUser,
@@ -159,7 +169,12 @@ exports.postRw = async (req,res) => {
             throw new Error("User belum terdaftar sebagai warga");
         }
         await WargaModel.findOneAndDelete({user: idUser});
-        
+
+        const cekRt = await rtModel.findOne({user: idUser});
+        if (cekRt) {
+            await rtModel.findOneAndDelete({user: idUser});
+        }
+
         const dataUser = await userModel.findByIdAndUpdate(idUser, {
             role: 3 // (rw)
         }, {new: true});
@@ -214,11 +229,23 @@ exports.postPerangkatDesa = async (req,res) => {
 
 
 ////// PIMPINAN DESA //////
-
 exports.postPimpinanDesa = async (req,res) => {
     try{
         const {idUser} = req.params;
         const {RolePemimpinDesa} = req.body;
+
+
+        const cekWarga = await WargaModel.findOne({user: idUser});
+        if (!cekWarga) {
+            throw new Error("User belum terdaftar sebagai warga");
+        }
+        const cekPimpinanDesa = await ppModel.findOne({rolePemimpinDesa: RolePemimpinDesa});
+        // jika pimpinan desa sudah ada yang memiliki role yang sama
+        if (cekPimpinanDesa) {
+            throw new Error("Pimpinan Desa sudah ada anda jangan membuat duplikat role dan melengserkan pimpinan desa yang sudah ada , anjay :)");
+        }
+        await WargaModel.findOneAndDelete({user: idUser});
+
         const dataUser = await userModel.findByIdAndUpdate(idUser, {
             role: 5, // (pimpinan desa)
